fix: fixed graph bug not re-arranging elements after first generation
